associate entity

+++ database/migrations/2021_06_08_050538_create_klients_table.php
@@ -17,7 +17,7 @@ class CreateKlientsTable extends Migration
             $table->id();
             $table->string('imie');
             $table->string('adres');
-            $table->integer('pesel');
+            $table->bigInteger('pesel');
             $table->timestamps();
         });
     }
